sqlite & exceptions
Major SQLite changes: now instead of returning false or error code (which we almost never checked anyway) on failures
exceptions will be thrown. Currently they are handled during db opening (due all the shenanigans we do there)
and plugin settings instantiation (for obvious reasons), i.e. SCTL_CREATE, as promised, will return FALSE in case of failure.
This should also fix 0002940: SCTL_CREATE: issues when creating multiple instances of settings.
In the rest of the code these possible exceptions are (currently) not handled, so you will notice if anything goes wrong.
SQLITE_BUSY is processed separately - after a 5-sec timeout it will be possible to either retry or abort the operation

In other words - these changes are quite experimental, here be dragons.

+++ far/editcontrol.cpp
@@ -594,7 +594,7 @@ int EditControl::AutoCompleteProc(bool Manual,bool DelBlock,Manager::Key& BackKe
 									if(ComplMenu->size() > 1)
 									{
 										const auto Data = ComplMenu->GetComplexUserDataPtr<cmp_user_data>();
-										if(Data && pHistory->DeleteIfUnlocked(Data->HistoryRecordId))
+										if(Data && Data->HistoryRecordId && pHistory->DeleteIfUnlocked(Data->HistoryRecordId))
 										{
 											ComplMenu->DeleteItem(ComplMenu->GetSelectPos());
 											if(ComplMenu->size() > 1)
